@@ -1,11 +1,18 @@
-import { HttpClientModule } from '@angular/common/http';
 import { NgModule } from '@angular/core';
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-//import { BrowserModule } from '@angular/platform-browser';
-import { BrowserModule } from '@angular/platform-browser';
-import { AppRoutingModule } from './app-routing.module';
+import { FormsModule } from '@angular/forms'
+import { ReactiveFormsModule } from '@angular/forms'
 import { AppComponent } from './app.component';
 import { MainComponent } from './header/main.component';
+import { CollapseModule } from 'ngx-bootstrap/collapse';
+import { BrowserModule } from '@angular/platform-browser';
+import { HttpClientModule } from '@angular/common/http';
+import { AppRoutingModule } from './app-routing.module';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { NgScrollbarModule } from 'ngx-scrollbar';
+import { BsDropdownModule } from 'ngx-bootstrap/dropdown';
+
+
+
 
 @NgModule({
   declarations: [
@@ -13,11 +20,15 @@ import { MainComponent } from './header/main.component';
     MainComponent    
   ],
   imports: [
+    CollapseModule,
     BrowserModule,
-    AppRoutingModule,
     FormsModule,
     ReactiveFormsModule,
-    HttpClientModule
+    HttpClientModule,
+    AppRoutingModule,
+    BrowserAnimationsModule,
+    NgScrollbarModule,
+    BsDropdownModule
   ],
   providers: [],
   bootstrap: [AppComponent]
